# openapi.yaml
openapi: 3.0.0
info:
  title: 꼬리콥터 API 명세
  description: 꼬리콥터 백엔드 API 명세서입니다. 초기 온보딩, 산책 기능, 코스 및 기록 관리를 위한 API들을 포함합니다.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: 개발 서버
...
paths:
...
  /locations/search:
    get:
      tags:
        - 인증 및 온보딩
      summary: 위치 검색
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 검색 결과 목록 반환
...
  /breeds/search:
    get:
      tags:
        - 인증 및 온보딩
      summary: 견종 검색
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 견종 검색 결과 목록 반환
...
  /marking-photos:
    post:
      tags:
        - 산책 기능
      summary: 새로운 마킹 포인트 등록 (50m 반경 통합 로직)
      requestBody:
        description: 마킹 포인트 사진 정보
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walkRecordId:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                photoUrl:
                  type: string
      responses:
        '201':
          description: 마킹 포인트 등록 성공
...
  /marking-photozones/{photozoneId}/photos:
    post:
      tags:
        - 산책 기능
      summary: 기존 포토존에 사진 추가
      parameters:
        - name: photozoneId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 추가할 사진 URL
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photoUrl:
                  type: string
                walkRecordId:
                  type: string
      responses:
        '201':
          description: 사진 추가 성공
...